requirements.txt
alembic==1.11.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
Werkzeug==2.3.4
zipp==3.15.0
